WIP: Modularize neovim settings
This commit prepares a full modularization of the neovim configuration.
In a first step, the content of the `.vimrc` file is split into multiple
lua files. On file is created per "domain", where the domains are chosen
rather arbitrarily for the moment. Everything is put into LUA files,
although most files simply call the old vimscript code using `vim.cmd`.